Ended session redirects to login page with proper user feedback, Sleep input data changed from integer to double

+++ src/main/java/com/racetrack/controller/LogApiController.java
@@ -294,7 +294,7 @@ public class LogApiController {
             Double mileage,
             Boolean hurting,
             String painDetails,
-            Integer sleepHours,
+            Double sleepHours,
             Integer stressLevel,
             Boolean plateProportion,
             Boolean gotThatBread,
